int/compiler/builder/propagated/directory: moves package
