+++ internal/compiler/builder/propagated/directory/inspect.go
