+++ internal/compiler/builder/propagated/directory/inspect_test.go
