+++ internal/compiler/builder/propagated/directory/meta.go
