+++ internal/compiler/builder/propagated/directory/testdata/acme/docs/magnet.md
